Move modelStubWithSingleGroup out of tests.
